Fixes serialPort thread issue.
At least has to solve.
The problem is that in Qt you can not work with serialPort
from different threads. You can work with it only in thread
where you created that. So I create separate thread which
holds queue of all requests to the port. Possibly it also can
solve the problem when robot does not go to the Direct Control
mode in old program. I think it may be connected with the fact that
port on computers are virtual and actually they work with one device.
And when you are trying to turn on DC mode on two robots in parallel
it may lead to conflicts. This treat can fix it too.